requirements.txt
numpy
pandas
pillow
matplotlib
learn2learn
opencv-python
line_profiler_pycharm